requirements.txt
networkx==3.3
numpy==1.26.4
pymongo==4.6.3
python-dotenv==1.0.1
Requests==2.31.0
tqdm==4.66.2
